+++ src/schemas/scripts/generateTypes.ts
@@ -1,12 +1,13 @@
 import fs from 'fs'
-import path from 'path'
 import { compile } from 'json-schema-to-typescript'
+import path from 'path'
 import { schemas } from '..'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const prettierConfig = require('@joystream/prettier-config')
 
 Object.entries(schemas).forEach(([schemaKey, schema]) => {
+  // eslint-disable-next-line @typescript-eslint/no-floating-promises
   compile(schema, `${schemaKey}Json`, {
     style: prettierConfig,
     ignoreMinAndMaxItems: true,